RAG_Chat_Bot/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
langchain
langchain-community
langchain-openai
faiss-cpu
pydantic
